6 -> clean code task

+++ routes/api.php
@@ -3,62 +3,52 @@
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Route;
 
-/*
-|--------------------------------------------------------------------------
-| API Routes
-|--------------------------------------------------------------------------
-|
-| Here is where you can register API routes for your application. These
-| routes are loaded by the RouteServiceProvider and all of them will
-| be assigned to the "api" middleware group. Make something great!
-|
-*/
-
-Route::middleware('auth:sanctum')->get('/user', function (Request $request) {
-    return $request->user();
-});
 
 Route::group(['namespace' => 'Api', 'prefix' => 'v1'], function () {
-    Route::post('/register', [\App\Http\Controllers\Api\v1\AuthController::class, 'register']);
-    Route::post('/login', [\App\Http\Controllers\Api\v1\AuthController::class, 'login']);
+    Route::post('/register', [\App\Http\Controllers\Api\v1\AuthController::class, 'register'])->name('register');
+    Route::post('/login', [\App\Http\Controllers\Api\v1\AuthController::class, 'login'])->name('login');
     Route::middleware('auth:api')
-        ->get('/logout', [\App\Http\Controllers\Api\v1\AuthController::class, 'logout']);
+        ->get('/logout', [\App\Http\Controllers\Api\v1\AuthController::class, 'logout'])->name('logout');
 });
 Route::group(['namespace' => 'Api', 'prefix' => 'v1'], function () {
-//    Route::middleware('auth:api')
-//        ->get('/rols', [\App\Http\Controllers\Api\v1\AuthController::class, 'logout']);
-//        ->get('/roles', function () {
-    Route::get('/roles', [\App\Http\Controllers\Api\v1\RoleController::class, 'index']);
-    Route::post('/roles', [\App\Http\Controllers\Api\v1\RoleController::class, 'store']);
-    Route::get('/roles/{id}', [\App\Http\Controllers\Api\v1\RoleController::class, 'show']);
-    Route::put('/roles/{id}', [\App\Http\Controllers\Api\v1\RoleController::class, 'update']);
-    Route::delete('/roles/{id}', [\App\Http\Controllers\Api\v1\RoleController::class, 'destroy']);
-    Route::get('/roles/{id}/users', [\App\Http\Controllers\Api\v1\RoleController::class, 'allUser']);
-    Route::post('/roles/{id}/users', [\App\Http\Controllers\Api\v1\RoleController::class, 'attachRole']);
-    Route::delete('/roles/{id}/users', [\App\Http\Controllers\Api\v1\RoleController::class, 'detachRole']);
+    Route::group(['middleware' => ['CheckRole','auth:api']], function () {
 
-    Route::get('/users', [\App\Http\Controllers\Api\v1\userController::class, 'index']);
-    Route::get('/users/{id}', [\App\Http\Controllers\Api\v1\userController::class, 'show']);
-    Route::delete('/users/{id}', [\App\Http\Controllers\Api\v1\userController::class, 'destroy']);
-    Route::get('/users/{id}/roles', [\App\Http\Controllers\Api\v1\userController::class, 'allRoles']);
-    Route::post('/users/{id}/roles', [\App\Http\Controllers\Api\v1\userController::class, 'attachRole']);
-    Route::delete('/users/{id}/roles', [\App\Http\Controllers\Api\v1\userController::class, 'detachRole']);
+        Route::get('/roles', [\App\Http\Controllers\Api\v1\RoleController::class, 'index'])->name('role_index');
+        Route::post('/roles', [\App\Http\Controllers\Api\v1\RoleController::class, 'store'])->name('role_store');
+        Route::get('/roles/{id}', [\App\Http\Controllers\Api\v1\RoleController::class, 'show'])->name('role_show');
+        Route::put('/roles/{id}', [\App\Http\Controllers\Api\v1\RoleController::class, 'update'])->name('role_update');
+        Route::delete('/roles/{id}', [\App\Http\Controllers\Api\v1\RoleController::class, 'destroy'])->name('role_destroy');
+        Route::get('/roles/{id}/users', [\App\Http\Controllers\Api\v1\RoleController::class, 'allUser'])->name('role_allUser');
+        Route::post('/roles/{id}/users', [\App\Http\Controllers\Api\v1\RoleController::class, 'attachRole'])->name('role_attachRole');
+        Route::delete('/roles/{id}/users', [\App\Http\Controllers\Api\v1\RoleController::class, 'detachRole'])->name('role_detachRole');
+
+        Route::get('/users', [\App\Http\Controllers\Api\v1\userController::class, 'index'])->name('user_index');
+        Route::get('/users/{id}', [\App\Http\Controllers\Api\v1\userController::class, 'show'])->name('user_show');
+        Route::delete('/users/{id}', [\App\Http\Controllers\Api\v1\userController::class, 'destroy'])->name('user_destroy');
+        Route::get('/users/{id}/roles', [\App\Http\Controllers\Api\v1\userController::class, 'allRoles'])->name('user_allRoles');
+        Route::post('/users/{id}/roles', [\App\Http\Controllers\Api\v1\userController::class, 'attachRole'])->name('user_attachRole');
+        Route::delete('/users/{id}/roles', [\App\Http\Controllers\Api\v1\userController::class, 'detachRole'])->name('user_detachRole');
+
+        Route::group(['prefix' => 'tasks'], function () {
+            Route::get('/', [\App\Http\Controllers\Api\v1\TaskController::class, 'index'])->name('task_index');
+            Route::post('/', [\App\Http\Controllers\Api\v1\TaskController::class, 'store'])->name('task_store');
+            Route::get('/{id}', [\App\Http\Controllers\Api\v1\TaskController::class, 'show'])->name('task_show');
+            Route::put('/{id}', [\App\Http\Controllers\Api\v1\TaskController::class, 'update'])->name('task_update');
+            Route::delete('/{id}', [\App\Http\Controllers\Api\v1\TaskController::class, 'destroy'])->name('task_destroy');
+
+        });
+
+        Route::get('/test2', function (Request $request) {
+//            dd($request);
+            return $request;
+        });
+    });
 
 });
 Route::group(['namespace' => 'Api', 'prefix' => 'v1'], function () {
-//    Route::middleware('auth:api')
-//        ->get('/rols', [\App\Http\Controllers\Api\v1\AuthController::class, 'logout']);
-//        ->get('/roles', function () {
-
-//    Route::get('/role_user', [\App\Http\Controllers\Api\v1\RoleUserController::class, 'index']);
-//    Route::post('/role_user', [\App\Http\Controllers\Api\v1\RoleUserController::class, 'store']);
-//    Route::get('/role_user/{id}', [\App\Http\Controllers\Api\v1\RoleUserController::class, 'show']);
-//    Route::put('/role_user/{id}', [\App\Http\Controllers\Api\v1\RoleUserController::class, 'update']);
-//    Route::delete('/role_user/{id}', [\App\Http\Controllers\Api\v1\RoleUserController::class, 'destroy']);
 });
 
 
 Route::get('/test', function (Request $request) {
-//    return "000122";
     return config('roles.models.role');
 });
